Reduced total memory size of ClockTime

+++ ClockTime.hpp
@@ -1,12 +1,14 @@
 #ifndef CLOCKTIME_H
 #define CLOCKTIME_H
 
+#include <Arduino.h>
+
 struct ClockTime
 {
-    unsigned long hours;
-    unsigned long mins;
-    unsigned long secs;
-    unsigned long ms;
+    uint8_t hours;
+    uint8_t mins;
+    uint8_t secs;
+    uint16_t ms;
 };
 
 ClockTime msToClockTime(unsigned long ms);
